+++ backend/src/controllers/crop.controller.js
@@ -19,8 +19,8 @@ export const getAllCrops = async(req, res)=>{
 export const getCropById = async (req,res) => {
     const {id} = req.params;
     try {
-        if(!id){
-            msg:"el id es invalido coloque un id valido"
+         if(!id){
+            return res.status(400).json({msg:"el id es invalido coloque un id valido"})
         }
         const crop = await Crop.findById(id).populate('farmer');
         return res.status(200).json({
@@ -61,7 +61,7 @@ export const updateCrop = async (req,res) => {
     const {name,state,amount,farmer} = req.body;
     try {
          if(!id){
-            msg:"el id es invalido coloque un id valido"
+            return res.status(400).json({msg:"el id es invalido coloque un id valido"})
         }
         const crop = await Crop.findByIdAndUpdate(id,
             {name,state,amount,farmer},
@@ -83,6 +83,9 @@ export const updateCrop = async (req,res) => {
 export const deleteCrop = async (req, res)=>{
     const {id} = req.params;
     try {
+         if(!id){
+            return res.status(400).json({msg:"el id es invalido coloque un id valido"})
+        }
         const crop = await Crop.findByIdAndDelete(id);
         return res.status(204).json({ msg: "Cultivo eliminado correctamente",data:crop });
     } catch (error) {
